requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
beautifulsoup4==4.13.5
billiard==4.2.1
bleach==6.2.0
cachetools==6.2.0
celery==5.5.3
cffi==2.0.0
channels==4.3.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
crispy-tailwind==1.0.3
cryptography==45.0.7
daphne==4.2.1
decorator==5.2.1
diff-match-patch==20241021
Django==5.2.6
django-admin-interface==0.30.1
django-appconf==1.1.0
django-auditlog==3.2.1
django-ckeditor==6.7.3
django-cleanup==9.0.0
django-colorfield==0.14.0
django-compressor==4.5.1
django-cors-headers==4.8.0
django-crispy-forms==2.4
django-csp==4.0
django-debug-toolbar==6.0.0
django-etc==1.4.0
django-extensions==4.1
django-extra-settings==0.14.1
django-filebrowser==4.0.3
django-filter==25.1
django-grappelli==3.0.10
django-guardian==3.1.3
django-guid==3.5.2
django-hitcount==1.3.5
django-import-export==4.3.9
django-jalali-date==2.0.0
django-js-asset==3.1.2
django-meta==2.5.0
django-model-utils==5.0.0
django-pwa==2.0.1
django-redirects==0.9.0
django-redis==6.0.0
django-storages==1.14.6
django-taggit==6.1.0
django-treebeard==4.7.1
django-unfold==0.65.0
django-unicorn==0.62.0
django_components==0.141.5
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djc_core_html_parser==1.0.2
drf-api-logger==1.1.21
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.9.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
iniconfig==2.1.0
jalali_core==1.0.0
jdatetime==5.2.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
Markdown==3.9
orjson==3.11.3
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.1.0
pytest==8.4.2
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-fsutil==0.15.0
python-slugify==8.0.4
PyYAML==6.0.2
rcssmin==1.1.2
redis==5.2.1
referencing==0.36.2
rjsmin==1.2.2
rpds-py==0.27.1
service-identity==24.2.0
setuptools==80.9.0
shortuuid==1.0.13
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
tablib==3.8.0
text-unidecode==1.3
Twisted==25.5.0
txaio==25.6.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
zope.interface==7.2
